Extend sequence number of the fragments map to 32 bits
Because the receiver adds an offset to those sequence numbers which
cross the 0xfff boundary, the key must be extended to 32 bits so
the frames are reconstructible

+++ include/formats/h265.hh
@@ -51,7 +51,7 @@ namespace uvg_rtp {
 
             /* map of frame's fragments,
              * allows out-of-order insertion and loop-through in order */
-            std::map<uint16_t, uvg_rtp::frame::rtp_frame *> fragments;
+            std::map<uint32_t, uvg_rtp::frame::rtp_frame *> fragments;
 
             /* storage for fragments that require relocation */
             std::vector<uvg_rtp::frame::rtp_frame *> temporary;
